working, needs clean up and further verification
