+++ src/data/modelmap.json
@@ -1,4 +1,4 @@
 {
-    "ai/lama3.3": "anthropic.claude-3-7-sonnet-20250219-v1:0",
-    "ai/gemma3": "google/gemini-2.5-pro-preview-03-25"
+    "ai/claude35": "anthropic.claude-3-5-sonnet-20241022-v2:0",
+    "ai/gemini25": "gemini-2.5-pro-preview-03-25"
 }
